Kernel: Make Driver interface more general
Now these are not USB specific, any one can register a driver class,
which has to implement an `init` function.
